Wokring GitHub actions; fixed formatting in readme; Altered a few file names; replace Class Libraries with NET Maui Class libraries
