+++ InvoiceGenerator-dotnet-maui-UI/ClientDetailsViewPage.xaml.cs
@@ -1,4 +1,4 @@
-using InvoiceGenerator.BusinessLogic;
+using InvoiceGenerator.Maui.BusinessLogic;
 using InvoiceGenerator_dotnet_maui_UI.ViewModels;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
