+++ InvoiceGenerator.Maui.BusinessLogic/IClientService.cs
@@ -1,0 +1,10 @@
+﻿using InvoiceGenerator.Maui.Domain.Models;
+
+namespace InvoiceGenerator.Maui.BusinessLogic
+{
+    public interface IClientService
+    {
+        List<Client> GetClientNames();
+        List<Client> GetClients();
+    }
+}
