+++ InvoiceGenerator.Maui.BusinessLogic/Platforms/iOS/PlatformClass1.cs
@@ -1,0 +1,7 @@
+﻿namespace InvoiceGenerator.Maui.BusinessLogic
+{
+    // All the code in this file is only included on iOS.
+    public class PlatformClass1
+    {
+    }
+}
