+++ InvoiceGenerator.Maui.Domain/Platforms/Tizen/PlatformClass1.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace InvoiceGenerator.Maui.Domain
+{
+    // All the code in this file is only included on Tizen.
+    public class PlatformClass1
+    {
+    }
+}
