+++ InvoiceGenerator.Maui.Domain/Platforms/iOS/PlatformClass1.cs
@@ -1,0 +1,7 @@
+﻿namespace InvoiceGenerator.Maui.Domain
+{
+    // All the code in this file is only included on iOS.
+    public class PlatformClass1
+    {
+    }
+}
