+++ readme.md
@@ -1,18 +1,24 @@
 # invoice-generator .NET Maui
+
 Invoice Generator but with .NET Maui.
 
 ## Building
+
 - Visual Studio 2022 Preview
+
 Currently, for .NET Maui Visual Studio 2022 Preview is required.
 
 Link here: [Visual Studio 2022 Preview](https://visualstudio.microsoft.com/vs/preview/#download-preview)
+
 - Setup Android Emulator
 - Enable HyperV Support
 - Enable Windows Developer Mode
 - Setup MacBuild Host for iOS Development
 
 ### Notes
+
 #### Gerald Verslius .NET Maui Crash Course
+
 Comprehensive crash course created by [Gerald Verslius](https://www.youtube.com/c/GeraldVersluis).
 
 Link below contains the YouTube Playlist and the GitHub Repository.
@@ -20,11 +26,13 @@ Link below contains the YouTube Playlist and the GitHub Repository.
 Link here: [Gerald Verslius Crash Course](https://blog.verslu.is/maui/dotnet-maui-crash-course/)
 
 #### James Montemagno .NET Maui Crash Course
+
 More in-depth crash course about .NET Maui created by [James Montemagno](https://www.youtube.com/c/JamesMontemagno).
 
 Link here: [James Montemagno Crash Course](https://www.youtube.com/watch?v=DuNLR_NJv8U)
 
 #### Microsoft .NET Maui Documentation
+
 Link here: [What is .NET Maui?](https://docs.microsoft.com/en-us/dotnet/maui/what-is-maui)
 
 Link here: [.NET Maui Controls](https://docs.microsoft.com/en-us/dotnet/maui/user-interface/controls/)
@@ -32,21 +40,27 @@ Link here: [.NET Maui Controls](https://docs.microsoft.com/en-us/dotnet/maui/use
 Link here: [In its Entirety](https://docs.microsoft.com/en-us/dotnet/maui/)
 
 #### .NET Maui Community Toolkit
+
 Link here: [Community Toolkit](https://codetraveler.io/2021/10/18/using-net-maui-community-toolkit-2/)
 
 #### Microsoft XAML Documentation
+
 Link here: [XAML Overview](https://docs.microsoft.com/en-us/dotnet/desktop/wpf/xaml/?view=netdesktop-6.0)
 
 #### dotNET Xamarin 101
+
 Link here: [Xamarin for Beginners](https://www.youtube.com/playlist?list=PLdo4fOcmZ0oU10SXt2W58pu2L0v2dOW-1)
 
 #### James Montemagno XAML 101
+
 Link here: [XAML 101 + Hello World!](https://www.youtube.com/watch?v=GLfR2uosoSw)
 
 #### Xamarin Forms Tutorial
+
 Link here: [Tutorial](https://www.youtube.com/playlist?list=PL-_mV1UKHGpRcFC834CsuVGEhZHjsUNIF)
 
 #### Telerik .NET Maui Controls
+
 Caution: Telerik is paid for, however, a free trial is possible
 
 Link here: [Telerik UI Controls](https://www.telerik.com/maui-ui)
